Rename `forks` module to indicate that it's for testing
